fix: Grandes mudanças.

+++ src/infra/repositories/user_repository.py
@@ -19,22 +19,28 @@ class UserRepository:
         user_model = self.db.query(User).filter(User.id == user_id).first()
         return user_model
 
+    def select_user_by_email(self, email: str):
+        user_model = self.db.query(User).filter(User.email == email).first()
+        return user_model
+
     def select_all(self):
         users_models = self.db.query(User).all()
         return users_models
 
     def update(self, user_id: int, user: User):
         user_model = self.db.query(User).filter(User.id == user_id).first()
-        user_model.name = user.name
-        user_model.email = user.email
-        user_model.telephone = user.telephone
-        user_model.password = user.password
-        self.db.commit()
-        self.db.refresh(user_model)
+        if user_model and user:
+            user_model.name = user.name
+            user_model.email = user.email
+            user_model.telephone = user.telephone
+            user_model.password = user.password
+            self.db.commit()
+            self.db.refresh(user_model)
         return user_model
 
     def delete(self, user_id: int):
         user_model = self.db.query(User).filter(User.id == user_id).first()
-        self.db.delete(user_model)
-        self.db.commit()
+        if user_model:
+            self.db.delete(user_model)
+            self.db.commit()
         return user_model
